+++ lib/src/server/_server.dart
@@ -6,7 +6,7 @@ export 'package:json_api/src/server/numbered_page.dart';
 export 'package:json_api/src/server/page.dart';
 export 'package:json_api/src/server/page_parameters.dart';
 export 'package:json_api/src/server/pagination_strategy.dart';
-export 'package:json_api/src/server/request/request.dart';
+export 'package:json_api/src/server/request.dart';
 export 'package:json_api/src/server/response.dart';
 export 'package:json_api/src/server/routing.dart';
 export 'package:json_api/src/server/server.dart';