+++ lib/src/server/response.dart
@@ -1,6 +1,6 @@
 import 'package:json_api/src/document/document.dart';
-import 'package:json_api/src/document/error.dart';
 import 'package:json_api/src/document/identifier.dart';
+import 'package:json_api/src/document/json_api_error.dart';
 import 'package:json_api/src/document/primary_data.dart';
 import 'package:json_api/src/document/resource.dart';
 import 'package:json_api/src/server/collection.dart';