+++ test/unit/document/resource_data_test.dart
@@ -5,7 +5,7 @@ import 'helper.dart';
 
 void main() {
   test('Can decode a primary resource with missing id', () {
-    final data = ResourceData.fromJson(recodeJson({
+    final data = ResourceData.decodeJson(recodeJson({
       'data': {'type': 'apples'}
     }));
     expect(data.toResource().type, 'apples');
@@ -13,7 +13,7 @@ void main() {
   });
 
   test('Can decode a primary resource with null id', () {
-    final data = ResourceData.fromJson(recodeJson({
+    final data = ResourceData.decodeJson(recodeJson({
       'data': {'type': 'apples', 'id': null}
     }));
     expect(data.toResource().type, 'apples');